Refactored and added or updated documentation - Moved some useful code to actions sections - Updated documentation for new functionality added to old components
