+++ components/Atoms/Button.tsx
@@ -12,11 +12,15 @@ interface ButtonProps {
 }
 
 /**
- * Displays a button.
+ * Displays a button which can be clicked to execute an action.
  * This button has many different variants:
  * - filled: a filled button with a background color
  * - outlined: a button with a border and no background color
  * - ghost: a button with no background color and no border
+ *
+ * The onClick prop can be a string or a function.
+ * If it is a string, it will redirect to the page.
+ * If it is a function, it will execute the function.
  *
  * @param onClick (string | function): the action to be performed when the button is clicked
  * @param variant (ButtonVariant): the variant of the button
